8/4 - Feedback from MK
Breakpoints/scaling:
The breakpoint issue has been resolved by adding a max-width to the
main container.

Deletions:
The facebook comment script has been removed, and the nav bar has
largely been commented out for the time being.

Font:
One font, Source Sans Pro, is used throughout our design (including
bold and italic versions of it). I can swap it out for the font in the
comps if you know what font it is. The size of the font for the
celebrity's name is going to present problems in the sense that it will
spill onto two lines most of the time (since most names are longer than
"Tom Hanks"). If you are okay with the name spanning two lines of text
we can leave the font size/weight as it is.

Margin/padding/spacing:
The margin and padding have been updated to be closer to the comps.
Regarding the quote bubble text (the quotes with colored backgrounds),
the placeholder quote happens to only take up a small portion of the
last line, and the design comps put the circular photo too far up into
the quote bubble. In the comps, if the last line of text in the bubble
extended to the end, the text would run into the photo. In order to
address this issue, the spacing has been adjusted based on a full line
of text at the end (screenshot below).

https://www.dropbox.com/s/mu8gyga62qfk1fr/ii_15dbe7d82933043f.png?dl=0