Convert to Storybook (and make it more compatible with the rest of our codebases) (#30)
This PR converts hollowverse/design to a Storybook-based website. This will allow us to more easily view the different states of our app in different viewports.

I've deployed it at https://hollowverse.github.io/design/ (and from now on any merge to master will automatically be deployed to this website)

This PR also brings this codebase closer to the common Hollowverse codebase conventions. Hollowverse conventions are not better or worse than the conventions that were in this codebase, it's just good to have consistency.

CC @Maxgrb 

Closes https://github.com/hollowverse/hollowverse/issues/498